:recycle: file 분리 및 리팩토링
