+++ RockPaperScissors/RockPaperScissors/HandType.swift
@@ -1,0 +1,14 @@
+//
+//  HandType.swift
+//  RockPaperScissors
+//
+//  Created by jin on 8/23/22.
+//
+
+import Foundation
+
+enum HandType: Int {
+    case scissor = 1
+    case rock = 2
+    case paper = 3
+}
