+++ RockPaperScissors/RockPaperScissors/Turn.swift
@@ -1,0 +1,13 @@
+//
+//  Turn.swift
+//  RockPaperScissors
+//
+//  Created by jin on 8/23/22.
+//
+
+import Foundation
+
+enum Turn: String {
+    case user = "사용자"
+    case computer = "컴퓨터"
+}
